rapidapi_mcp_server/requirements.txt
fastmcp>=0.1.0
aiohttp>=3.9.0
python-dotenv>=1.0.0 